Move Task struct to imgmanager
