+++ internal/imgmanager/task.go
@@ -1,0 +1,14 @@
+package imgmanager
+
+import (
+	corev1 "k8s.io/api/core/v1"
+
+	ofenv1 "github.com/cybozu-go/ofen/api/v1"
+)
+
+type Task struct {
+	Ref              string
+	RegistryPolicy   ofenv1.RegistryPolicy
+	NodeImageSetName string
+	Secrets          *[]corev1.Secret
+}
